new bits in bitmask 32. bitwise_mask_ctrl.sh: simple interface for bitmask, It doesnt work yet and hasnt been tested. main.c: functions dev_write() and dev_read() is working. proc.c It does worked well, need revision. It is first working wersion

+++ externs.h
@@ -5,7 +5,7 @@
 extern const struct proc_ops byte_conv_mask_ops;
 extern const char * const byte_conv_dev_fname;
 extern const char * const byte_conv_proc_fname;
-extern u16 byte_conv_mask;
+extern u32 byte_conv_mask;
 extern spinlock_t byte_conv_mask_lock;
 
 #endif // EXTERNS_H__
